pyproject.toml
[project]
name = "entrenamiento-hibrido-ocr-texto"
version = "0.1.0"
description = "Procesamiento híbrido de PDFs con OCR + extracción de texto"
requires-python = ">=3.10"
dependencies = [
    "paddleocr==2.8.1",
    "paddlepaddle-gpu==2.6.2",
    # Dependencias GPU para PaddlePaddle (CUDA 11 compatible)
    "nvidia-cudnn-cu11==8.9.5.29",
    "nvidia-cublas-cu11==11.11.3.6",
    "nvidia-cuda-nvrtc-cu11==11.8.89",
    # Resto de dependencias
    "opencv-python==4.10.0.84",
    "PyMuPDF==1.26.4",
    "numpy==1.26.4",
    "pillow==11.3.0",
    "regex==2025.9.18",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "pandas>=2.1.0",
    "streamlit>=1.38.0",
    "selenium>=4.0.0",
    "webdriver-manager>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
# Configuración específica de uv
dev-dependencies = []
package = false